third_party/WebKit/LayoutTests/virtual/threaded/inspector/tracing/idle-callback.html

<html>
<head>
<script src="../../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../../http/tests/inspector/timeline-test.js"></script>
<script>
function performActions(callback)
{
    var requestId = window.requestIdleCallback(idleCallback);
    window.cancelIdleCallback(requestId);
    window.requestIdleCallback(idleCallback);
    function idleCallback()
    {
        window.requestIdleCallback(slowIdleCallback);
    }
    function slowIdleCallback(deadline)
    {
        while (deadline.timeRemaining()) {};
        if (callback)
            callback();
    }
}

function test()
{
    InspectorTest.invokeAsyncWithTimeline("performActions", finish);

    function finish()
    {
        InspectorTest.printTimelineRecordsWithDetails("RequestIdleCallback");
        InspectorTest.printTimelineRecordsWithDetails("CancelIdleCallback");
        InspectorTest.printTimelineRecordsWithDetails("FireIdleCallback");
        InspectorTest.completeTest();
    }
}

if (!window.testRunner)
    setTimeout(performActions, 2000);
</script>
</head>

<body onload="runTest()">
<p>
Tests Timeline events emitted when idle callback is scheduled and fired.
</p>

</body>
</html>
